+++ Tests/NukeTests/ImagePipelineTests/ImagePipelineConfigurationTests.swift
@@ -29,58 +29,6 @@ class ImagePipelineConfigurationTests: XCTestCase {
         XCTAssertNotNil(pipeline.configuration.dataCache)
     }
 
-    // MARK: Changing Callback Queue
-
-    func testChangingCallbackQueueLoadImage() {
-        // Given
-        let queue = DispatchQueue(label: "testChangingCallbackQueue")
-        let queueKey = DispatchSpecificKey<Void>()
-        queue.setSpecific(key: queueKey, value: ())
-
-        let dataLoader = MockDataLoader()
-        let pipeline = ImagePipeline {
-            $0.dataLoader = dataLoader
-            $0.imageCache = nil
-
-            $0._callbackQueue = queue
-        }
-
-        // When/Then
-        let expectation = self.expectation(description: "Image Loaded")
-        pipeline.loadImage(with: Test.request, progress: { _, _, _ in
-            XCTAssertNotNil(DispatchQueue.getSpecific(key: queueKey))
-        }, completion: { _ in
-            XCTAssertNotNil(DispatchQueue.getSpecific(key: queueKey))
-            expectation.fulfill()
-        })
-        wait()
-    }
-
-    func testChangingCallbackQueueLoadData() {
-        // Given
-        let queue = DispatchQueue(label: "testChangingCallbackQueue")
-        let queueKey = DispatchSpecificKey<Void>()
-        queue.setSpecific(key: queueKey, value: ())
-
-        let dataLoader = MockDataLoader()
-        let pipeline = ImagePipeline {
-            $0.dataLoader = dataLoader
-            $0.imageCache = nil
-
-            $0._callbackQueue = queue
-        }
-
-        // When/Then
-        let expectation = self.expectation(description: "Image data Loaded")
-        pipeline.loadData(with: Test.request, progress: { _, _ in
-            XCTAssertNotNil(DispatchQueue.getSpecific(key: queueKey))
-        }, completion: { _ in
-            XCTAssertNotNil(DispatchQueue.getSpecific(key: queueKey))
-            expectation.fulfill()
-        })
-        wait()
-    }
-
     func testEnablingSignposts() {
         ImagePipeline.Configuration.isSignpostLoggingEnabled = false // Just padding
         ImagePipeline.Configuration.isSignpostLoggingEnabled = true