+++ Tests/NukeTests/ImageProcessorsTests/ResizeTests.swift
@@ -143,7 +143,7 @@ import UIKit
         // Given an image with `right` orientation. From the user perspective,
         // the image a landscape image with s size 640x480px. The raw pixel
         // data, on the other hand, is 480x640px.
-        let input = try #require(Test.image(named: "right-orientation.jpeg"))
+        let input = Test.image(named: "right-orientation.jpeg")
         #expect(input.imageOrientation == .right)
 
         // When we resize the image to fit 320x480px frame, we expect the processor
@@ -162,7 +162,7 @@ import UIKit
         // Given an image with `right` orientation. From the user perspective,
         // the image a landscape image with s size 640x480px. The raw pixel
         // data, on the other hand, is 480x640px.
-        let input = try #require(Test.image(named: "right-orientation.jpeg"))
+        let input = Test.image(named: "right-orientation.jpeg")
         #expect(input.imageOrientation == .right)
 
         // When